refactor: extract AvrdudeRunner module as injectable seam for testing
